+++ Universal/Sniffs/WhiteSpace/DisallowInlineTabsSniff.php
@@ -94,7 +94,12 @@ final class DisallowInlineTabsSniff implements Sniff
             $this->tabWidth = Helper::getTabWidth($phpcsFile);
         }
 
+        if (defined('PHP_CODESNIFFER_IN_TESTS')) {
+            $this->tabWidth = Helper::getCommandLineData($phpcsFile, 'tabWidth');
+        }
+
         $tokens = $phpcsFile->getTokens();
+        $dummy  = new DummyTokenizer('', $phpcsFile->config);
 
         for ($i = 0; $i < $phpcsFile->numTokens; $i++) {
             // Skip all non-whitespace tokens and skip whitespace at the start of a new line.
@@ -109,10 +114,18 @@ final class DisallowInlineTabsSniff implements Sniff
             // If tabs haven't been converted to spaces by the tokenizer, do so now.
             $token = $tokens[$i];
             if (isset($token['orig_content']) === false) {
-                $dummy = new DummyTokenizer('', $phpcsFile->config);
-                $dummy->replaceTabsInToken($token, ' ', ' ', $this->tabWidth);
+                if ($token['content'] === '' || \strpos($token['content'], "\t") === false) {
+                    // If there are no tabs, we can continue, no matter what.
+                    continue;
+                }
+
+                $dummy->replaceTabsInToken($token);
             }
 
+            /*
+             * Tokens only have the 'orig_content' key if they contain tabs,
+             * so from here on out, we **know** there will be tabs in the content.
+             */
             $origContent = $token['orig_content'];
 
             $multiLineComment = false;
@@ -131,8 +144,6 @@ final class DisallowInlineTabsSniff implements Sniff
                 if ($commentOnly === '' || \strpos($commentOnly, "\t") === false) {
                     continue;
                 }
-            } elseif ($origContent === '' || \strpos($origContent, "\t") === false) {
-                continue;
             }
 
             $fix = $phpcsFile->addFixableError(